2010-02-04 19:14 UTC+0100 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * harbour/include/hbapifs.h   * harbour/src/rtl/filebuf.c   * harbour/contrib/hbmemio/memio.c   * harbour/contrib/hbnetio/netiocli.c     + added new method:          void Flush( PHB_FILE pFile, HB_BOOL fDirty );       to replaceable RDD IO structure (HB_FILE)     + added new function:          void hb_fileFlush( PHB_FILE pFile, HB_BOOL fDirty );
  * harbour/src/rdd/dbfntx/dbfntx1.c
  * harbour/src/rdd/dbfnsx/dbfnsx1.c
  * harbour/src/rdd/dbfcdx/dbfcdx1.c
  * harbour/src/rdd/dbffpt/dbffpt1.c
    * use hb_fileFlush()

  * harbour/src/vm/hvm.c
  * harbour/include/hbxvm.h
    ! fixed hb_xvm{Local,Static,Module}Name() to use 'const char *'
      instead of 'char *' - modifications by Xavi (jarabal)

  * harbour/include/hbapi.h
  * harbour/include/hbxvm.h
  * harbour/include/hbmacro.h
  * harbour/include/hbcompdf.h
  * harbour/src/vm/hvm.c
  * harbour/src/vm/macro.c
  * harbour/src/vm/memvars.c
    * replaced some 'BYTE' types with 'unsigned char' or 'int' types
    * cleanup some other casting and types to reduce conversions
      between function calls

  * harbour/src/common/expropt1.c
    ; updated comments

  * harbour/include/hbpp.h
  * harbour/include/hbexprop.h
  * harbour/include/hbexpra.c
  * harbour/include/hbexprb.c
    * eliminated HB_EXPR_PCODE?() macros
    ! fixed potential GPF trap during compilation of _GET_() functions
      which can be exploited by strange code using some special expressions
      as get variable.
      TODO: eliminate hb_compExprClone() used in hb_compExprSetGetBlock(),
            our optimization module does not care about shared expressions
            so it can be source of serious problems in the future if someone
            use this functions for non optimized expression which will be
            reduced later.
    * emulate clipper behavior for codeblock optimizations and do
      not optimize all codeblocks if -kc switch is used and codeblocks
      is inside non optimized part of code like (Clipper has such
      places)
    % simplified and optimized some functions like hb_compExprReduceList()

+++ harbour/include/hbcompdf.h
@@ -531,7 +531,7 @@ typedef struct _HB_COMMON
 {
    /* common to macro compiler members */
    int    mode;               /* HB_MODE_* */
-   ULONG  supported;          /* various flags for supported capabilities */
+   int    supported;          /* various flags for supported capabilities */
    const struct _HB_COMP_FUNCS * funcs;
 }
 HB_COMMON, * HB_COMMON_PTR;
@@ -558,7 +558,7 @@ typedef struct HB_MACRO_    /* a macro compiled pcode container */
 {
    /* common to compiler members */
    int    mode;            /* HB_MODE_* */
-   ULONG  supported;       /* various flags for supported capabilities */
+   int    supported;       /* various flags for supported capabilities */
    const struct _HB_COMP_FUNCS * funcs;
 
    /* macro compiler only members */
@@ -635,7 +635,7 @@ typedef struct _HB_COMP
 {
    /* common to macro compiler members */
    int    mode;            /* HB_MODE_* */
-   ULONG  supported;       /* various flags for supported capabilities */
+   int    supported;       /* various flags for supported capabilities */
    const struct _HB_COMP_FUNCS * funcs;
 
    /* compiler only members */
